refactor(core): extract chain_params (block time, chain id, hash, window)
Pulled the small remaining "chain parameter" constants out of
blockchain.rs into crate::chain_params:

  BLOCK_TIME_SECS   — per-block timing floor
  MAX_TX_PER_BLOCK  — per-block tx cap
  CHAIN_ID + get_chain_id()  — chain identifier + env override
  HASH_VERSION      — hash-algorithm version marker
  CHAIN_WINDOW_SIZE — in-memory sliding-window cache size

blockchain.rs re-exports all six names so existing import paths
keep resolving bit-identically (`crate::blockchain::BLOCK_TIME_SECS`,
`crate::blockchain::get_chain_id`, etc.).

Only one cross-crate caller (`bin/sentrix/main.rs` imports
`BLOCK_TIME_SECS` via `sentrix::core::blockchain::BLOCK_TIME_SECS`) —
keeps working through the re-export.

Tests use compile-time `const _: () = assert!(...)` for invariants
that are pure constant arithmetic (catch silent drift between paired
constants at build time, not just `cargo test`). Plus two runtime
tests for `get_chain_id` covering default + env-override + invalid-
input fallback.

sentrix-core test suite: 230 passed (was 222 on main).

Sixth refactor PR in the blockchain.rs decomposition series. After
all the in-flight PRs land:
  blockchain.rs LOC: 3,967 → ~3,470 (-12.5%)
